requirements.txt
streamlit
requests
pypdf
python-docx
PyPDF2